requirements.txt
streamlit==1.32.0
yfinance==0.2.36
pandas==2.2.0
plotly==5.18.0
numpy==1.26.3
python-dotenv==1.0.1
requests==2.31.0